+++ cached_network_image/lib/src/image_provider/multi_image_stream_completer.dart
@@ -7,7 +7,7 @@ import 'package:flutter/scheduler.dart';
 
 /// Slows down animations by this factor to help in development.
 double get timeDilation => _timeDilation;
-double _timeDilation = 1.0;
+double _timeDilation = 1;
 
 /// An ImageStreamCompleter with support for loading multiple images.
 class MultiImageStreamCompleter extends ImageStreamCompleter {
@@ -23,6 +23,24 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
     ImageStreamListener? imageStreamListener,
   })  : _informationCollector = informationCollector,
         _scale = scale {
+    codec.listen(
+      (event) {
+        if (_timer != null) {
+          _nextImageCodec = event;
+        } else {
+          _handleCodecReady(event);
+        }
+      },
+      onError: (Object error, StackTrace stack) {
+        reportError(
+          context: ErrorDescription('resolving an image codec'),
+          exception: error,
+          stack: stack,
+          informationCollector: informationCollector,
+          silent: true,
+        );
+      },
+    );
     if(imageStreamListener != null) {
       addListener(imageStreamListener);
     }
@@ -42,9 +60,9 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
       );
     });
     if (chunkEvents != null) {
-      chunkEvents.listen(
+      _chunkSubscription = chunkEvents.listen(
         reportImageChunkEvent,
-        onError: (dynamic error, StackTrace stack) {
+        onError: (Object error, StackTrace stack) {
           reportError(
             context: ErrorDescription('loading an image'),
             exception: error,
@@ -62,16 +80,26 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
   final double _scale;
   final InformationCollector? _informationCollector;
   ui.FrameInfo? _nextFrame;
+
   // When the current was first shown.
   Duration? _shownTimestamp;
+
   // The requested duration for the current frame;
   Duration? _frameDuration;
+
   // How many frames have been emitted so far.
   int _framesEmitted = 0;
   Timer? _timer;
+  StreamSubscription<ImageChunkEvent>? _chunkSubscription;
 
   // Used to guard against registering multiple _handleAppFrame callbacks for the same frame.
   bool _frameCallbackScheduled = false;
+
+  /// We must avoid disposing a completer if it never had a listener, even
+  /// if all [keepAlive] handles get disposed.
+  bool __hadAtLeastOneListener = false;
+
+  bool __disposed = false;
 
   void _switchToNewCodec() {
     _framesEmitted = 0;
@@ -122,7 +150,7 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
   Future<void> _decodeNextFrameAndSchedule() async {
     try {
       _nextFrame = await _codec!.getNextFrame();
-    } catch (exception, stack) {
+    } on Object catch (exception, stack) {
       reportError(
         context: ErrorDescription('resolving an image frame'),
         exception: exception,
@@ -163,6 +191,7 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
 
   @override
   void addListener(ImageStreamListener listener) {
+    __hadAtLeastOneListener = true;
     if (!hasListeners && _codec != null) _decodeNextFrameAndSchedule();
     super.addListener(listener);
   }
@@ -173,6 +202,56 @@ class MultiImageStreamCompleter extends ImageStreamCompleter {
     if (!hasListeners) {
       _timer?.cancel();
       _timer = null;
+      __maybeDispose();
     }
+  }
+
+  int __keepAliveHandles = 0;
+
+  @override
+  ImageStreamCompleterHandle keepAlive() {
+    final delegateHandle = super.keepAlive();
+    return _MultiImageStreamCompleterHandle(this, delegateHandle);
+  }
+
+  void __maybeDispose() {
+    if (!__hadAtLeastOneListener ||
+        __disposed ||
+        hasListeners ||
+        __keepAliveHandles != 0) {
+      return;
+    }
+
+    __disposed = true;
+
+    _chunkSubscription?.onData(null);
+    _chunkSubscription?.cancel();
+    _chunkSubscription = null;
+  }
+}
+
+class _MultiImageStreamCompleterHandle implements ImageStreamCompleterHandle {
+  _MultiImageStreamCompleterHandle(this._completer, this._delegateHandle) {
+    _completer!.__keepAliveHandles += 1;
+  }
+
+  MultiImageStreamCompleter? _completer;
+  final ImageStreamCompleterHandle _delegateHandle;
+
+  /// Call this method to signal the [ImageStreamCompleter] that it can now be
+  /// disposed when its last listener drops.
+  ///
+  /// This method must only be called once per object.
+  @override
+  void dispose() {
+    assert(_completer != null);
+    assert(_completer!.__keepAliveHandles > 0);
+    assert(!_completer!.__disposed);
+
+    _delegateHandle.dispose();
+
+    _completer!.__keepAliveHandles -= 1;
+    _completer!.__maybeDispose();
+    _completer = null;
   }
 }