+++ cached_network_image_web/lib/cached_network_image_web.dart
@@ -1,35 +1,98 @@
+/// Web implementation of CachedNetworkImage
 library cached_network_image_web;
 
 import 'dart:async';
+import 'dart:typed_data';
 import 'dart:ui' as ui;
 
 import 'package:cached_network_image_platform_interface'
         '/cached_network_image_platform_interface.dart' as platform
-    show ImageLoader;
-import 'package:cached_network_image_platform_interface'
-        '/cached_network_image_platform_interface.dart'
-    show ImageRenderMethodForWeb;
+    show ImageLoader, ErrorListener, ImageRenderMethodForWeb;
 import 'package:flutter/material.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 
 /// ImageLoader class to load images on the web platform.
 class ImageLoader implements platform.ImageLoader {
+  @Deprecated('Use loadImageAsync instead')
   @override
-  Stream<ui.Codec> loadAsync(
+  Stream<ui.Codec> loadBufferAsync(
     String url,
     String? cacheKey,
     StreamController<ImageChunkEvent> chunkEvents,
-    DecoderCallback decode,
+    DecoderBufferCallback decode,
     BaseCacheManager cacheManager,
     int? maxHeight,
     int? maxWidth,
     Map<String, String>? headers,
-    Function()? errorListener,
-    ImageRenderMethodForWeb imageRenderMethodForWeb,
-    Function() evictImage,
+    VoidCallback? errorListener,
+    platform.ImageRenderMethodForWeb imageRenderMethodForWeb,
+    VoidCallback evictImage,
+  ) {
+    return _load(
+      url,
+      cacheKey,
+      chunkEvents,
+      (bytes) async {
+        final buffer = await ui.ImmutableBuffer.fromUint8List(bytes);
+        return decode(buffer);
+      },
+      cacheManager,
+      maxHeight,
+      maxWidth,
+      headers,
+      (_) {},
+      imageRenderMethodForWeb,
+      evictImage,
+    );
+  }
+
+  @override
+  Stream<ui.Codec> loadImageAsync(
+    String url,
+    String? cacheKey,
+    StreamController<ImageChunkEvent> chunkEvents,
+    ImageDecoderCallback decode,
+    BaseCacheManager cacheManager,
+    int? maxHeight,
+    int? maxWidth,
+    Map<String, String>? headers,
+    ValueChanged<Object>? errorListener,
+    platform.ImageRenderMethodForWeb imageRenderMethodForWeb,
+    VoidCallback evictImage,
+  ) {
+    return _load(
+      url,
+      cacheKey,
+      chunkEvents,
+      (bytes) async {
+        final buffer = await ui.ImmutableBuffer.fromUint8List(bytes);
+        return decode(buffer);
+      },
+      cacheManager,
+      maxHeight,
+      maxWidth,
+      headers,
+      errorListener,
+      imageRenderMethodForWeb,
+      evictImage,
+    );
+  }
+
+  Stream<ui.Codec> _load(
+    String url,
+    String? cacheKey,
+    StreamController<ImageChunkEvent> chunkEvents,
+    _FileDecoderCallback decode,
+    BaseCacheManager cacheManager,
+    int? maxHeight,
+    int? maxWidth,
+    Map<String, String>? headers,
+    platform.ErrorListener? errorListener,
+    platform.ImageRenderMethodForWeb imageRenderMethodForWeb,
+    VoidCallback evictImage,
   ) {
     switch (imageRenderMethodForWeb) {
-      case ImageRenderMethodForWeb.HttpGet:
+      case platform.ImageRenderMethodForWeb.HttpGet:
         return _loadAsyncHttpGet(
           url,
           cacheKey,
@@ -42,8 +105,8 @@ class ImageLoader implements platform.ImageLoader {
           errorListener,
           evictImage,
         );
-      case ImageRenderMethodForWeb.HtmlImage:
-        return _loadAsyncHtmlImage(url, chunkEvents, decode).asStream();
+      case platform.ImageRenderMethodForWeb.HtmlImage:
+        return _loadAsyncHtmlImage(url, chunkEvents).asStream();
     }
   }
 
@@ -51,52 +114,54 @@ class ImageLoader implements platform.ImageLoader {
     String url,
     String? cacheKey,
     StreamController<ImageChunkEvent> chunkEvents,
-    DecoderCallback decode,
+    _FileDecoderCallback decode,
     BaseCacheManager cacheManager,
     int? maxHeight,
     int? maxWidth,
     Map<String, String>? headers,
-    Function()? errorListener,
-    Function() evictImage,
+    platform.ErrorListener? errorListener,
+    VoidCallback evictImage,
   ) async* {
     try {
-      await for (var result in cacheManager.getFileStream(url,
-          withProgress: true, headers: headers)) {
+      await for (final result in cacheManager.getFileStream(
+        url,
+        key: cacheKey,
+        withProgress: true,
+        headers: headers,
+      )) {
         if (result is DownloadProgress) {
-          chunkEvents.add(ImageChunkEvent(
-            cumulativeBytesLoaded: result.downloaded,
-            expectedTotalBytes: result.totalSize,
-          ));
+          chunkEvents.add(
+            ImageChunkEvent(
+              cumulativeBytesLoaded: result.downloaded,
+              expectedTotalBytes: result.totalSize,
+            ),
+          );
         }
         if (result is FileInfo) {
-          var file = result.file;
-          var bytes = await file.readAsBytes();
-          var decoded = await decode(bytes);
+          final file = result.file;
+          final bytes = await file.readAsBytes();
+          final decoded = await decode(bytes);
           yield decoded;
         }
       }
-    } catch (e) {
+    } on Object catch (e) {
       // Depending on where the exception was thrown, the image cache may not
       // have had a chance to track the key in the cache at all.
       // Schedule a microtask to give the cache a chance to add the key.
       scheduleMicrotask(() {
         evictImage();
       });
-
-      errorListener?.call();
+      errorListener?.call(e);
       rethrow;
-    } finally {
-      await chunkEvents.close();
     }
+    await chunkEvents.close();
   }
 
   Future<ui.Codec> _loadAsyncHtmlImage(
     String url,
     StreamController<ImageChunkEvent> chunkEvents,
-    DecoderCallback decode,
   ) {
     final resolved = Uri.base.resolve(url);
-
     // ignore: undefined_function
     return ui.webOnlyInstantiateImageCodecFromUrl(
       resolved,
@@ -111,3 +176,5 @@ class ImageLoader implements platform.ImageLoader {
     ) as Future<ui.Codec>;
   }
 }
+
+typedef _FileDecoderCallback = Future<ui.Codec> Function(Uint8List);